Update comments and rename some modules according to #37 review notes
