+++ zebra-consensus/src/lib.rs
@@ -66,4 +66,4 @@ pub use router::RouterError;
 /// A boxed [`std::error::Error`].
 pub type BoxError = Box<dyn std::error::Error + Send + Sync + 'static>;
 
-mod zsa;
+mod orchard_zsa;
