+++ zebrad/Cargo.toml
@@ -52,7 +52,6 @@ features = [
 
 [features]
 # In release builds, don't compile debug logging code, to improve performance.
-#default = ["release_max_level_info", "progress-bar", "getblocktemplate-rpcs"]
 default = ["release_max_level_info", "progress-bar", "getblocktemplate-rpcs", "tx-v6"]
 
 # Default features for official ZF binary release builds
